refactor(test): split calcs-coverage.js into per-calc files
One file per calc, named after the module it covers. Integration tests
(plugin index.js schema/uiSchema, plus the existing plugin.start stream-pipeline
suite) moved under test/integration/. Shared makeApp/makePlugin factories
extracted to test/helpers.js.

.mocharc.cjs added because mocha's default glob ('test/*.js') does not
recurse into subdirectories; stryker already uses 'test/**/*.js' so its
config is unchanged.

No behaviour change. BUG-pinning tests preserved verbatim for the fix
pass to flip in a follow-up commit.
