+++ test/courseOverGroundMagnetic.js
@@ -1,0 +1,35 @@
+const chai = require('chai')
+chai.Should()
+
+const { makeApp, makePlugin } = require('./helpers')
+
+describe('courseOverGroundMagnetic (extra branches)', () => {
+  const calc = require('../calcs/courseOverGroundMagnetic')
+
+  it('falls back to getSelfPath when magneticVariation is the 9999 sentinel', () => {
+    const app = makeApp({
+      selfPaths: { navigation: { magneticVariation: { value: 0.1 } } }
+    })
+    const d = calc(app, makePlugin())
+    const out = d.calculator(0.5, 9999)
+    out[0].value.should.be.closeTo(0.4, 1e-9)
+  })
+
+  it('computes a valid result', () => {
+    const d = calc(makeApp(), makePlugin())
+    const out = d.calculator(1.0, 0.1)
+    out[0].value.should.be.closeTo(0.9, 1e-9)
+  })
+
+  it('wraps negative results into [0, 2*PI)', () => {
+    const d = calc(makeApp(), makePlugin())
+    const out = d.calculator(0.1, 0.3)
+    out[0].value.should.be.closeTo(2 * Math.PI - 0.2, 1e-9)
+  })
+
+  it('wraps results above 2*PI back into [0, 2*PI)', () => {
+    const d = calc(makeApp(), makePlugin())
+    const out = d.calculator(6.28, -0.5)
+    out[0].value.should.be.closeTo(6.28 + 0.5 - 2 * Math.PI, 1e-9)
+  })
+})
