+++ test/windGround.js
@@ -1,0 +1,59 @@
+// Tests marked with `// BUG: ...` lock the CURRENT (incorrect) behaviour
+// of the module so the suite stays green today. A follow-up pass flips
+// those assertions to the correct behaviour and fixes the implementations.
+
+const chai = require('chai')
+chai.Should()
+
+const { makeApp, makePlugin } = require('./helpers')
+
+describe('windGround (extra branches)', () => {
+  const calcs = require('../calcs/windGround')
+
+  it('computes the expected values for a finite input vector', () => {
+    const arr = calcs(makeApp(), makePlugin())
+    const ground = arr[0]
+    const out = ground.calculator(1.0, 3.0, 5.0, 0.5)
+    out.should.have.lengthOf(3)
+    out[0].path.should.equal('environment.wind.directionTrue')
+    out[1].path.should.equal('environment.wind.angleTrueGround')
+    out[2].path.should.equal('environment.wind.speedOverGround')
+    out[0].value.should.be.closeTo(2.0459686742419585, 1e-9)
+    out[1].value.should.be.closeTo(1.0459686742419587, 1e-9)
+    out[2].value.should.be.closeTo(2.769931974487608, 1e-9)
+  })
+
+  // BUG: the path for ground-frame wind direction should be
+  // environment.wind.directionGround, but this calculator writes to
+  // environment.wind.directionTrue (water frame per the SK spec). The
+  // deprecated sister calc uses the correct directionGround path.
+  it('writes ground-frame direction to environment.wind.directionTrue (wrong path)', () => {
+    const arr = calcs(makeApp(), makePlugin())
+    const ground = arr[0]
+    const out = ground.calculator(1.0, 3.0, 5.0, 0.5)
+    out[0].path.should.equal('environment.wind.directionTrue')
+  })
+
+  it('uses awa as the angle when aws is effectively zero', () => {
+    const arr = calcs(makeApp(), makePlugin())
+    const ground = arr[0]
+    const out = ground.calculator(1.0, 3.0, 1e-12, 0.5)
+    out[1].value.should.be.closeTo(0.5, 1e-9)
+  })
+
+  it('deprecated calc returns null path when inputs are non-finite', () => {
+    const arr = calcs(makeApp(), makePlugin())
+    const deprecated = arr[1]
+    const out = deprecated.calculator(null, 0.1)
+    out.should.deep.equal([
+      { path: 'environment.wind.directionGround', value: null }
+    ])
+  })
+
+  it('deprecated calc returns a wrapped compass angle for finite inputs', () => {
+    const arr = calcs(makeApp(), makePlugin())
+    const deprecated = arr[1]
+    const out = deprecated.calculator(2, 1)
+    out[0].value.should.be.closeTo(3, 1e-9)
+  })
+})
